# openapi.yaml
openapi: 3.1.0
info:
  title: Smart City MVP Ticket System
  version: 1.0.0
paths:
...
  /api/tickets:
    post:
      tags:
      - tickets
      summary: Submit Ticket
      operationId: submit_ticket_api_tickets_post
      parameters:
      - name: X-Max-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Max-User-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitTicketIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - tickets
      summary: List Tickets
      operationId: list_tickets_api_tickets_get
      security:
      - HTTPBearer: []
      parameters:
      - name: role
        in: query
        required: true
        schema:
          type: string
          description: org|resident
          title: Role
        description: org|resident
      - name: X-Max-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Max-User-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketOut'
                title: Response List Tickets Api Tickets Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/tickets/{ticket_id}:
    get:
      tags:
      - tickets
      summary: Get Ticket
      operationId: get_ticket_api_tickets__ticket_id__get
      parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: string
          title: Ticket Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/tickets/{ticket_id}/accept:
    post:
      tags:
      - tickets
      summary: Accept Ticket
      operationId: accept_ticket_api_tickets__ticket_id__accept_post
      security:
      - HTTPBearer: []
      parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: string
          title: Ticket Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/tickets/{ticket_id}/assign:
    post:
      tags:
      - tickets
      summary: Assign Ticket
      operationId: assign_ticket_api_tickets__ticket_id__assign_post
      security:
      - HTTPBearer: []
      parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: string
          title: Ticket Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/tickets/{ticket_id}/complete:
    post:
      tags:
      - tickets
      summary: Complete Ticket
      operationId: complete_ticket_api_tickets__ticket_id__complete_post
      security:
      - HTTPBearer: []
      parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: string
          title: Ticket Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/tickets/{ticket_id}/cancel:
    post:
      tags:
      - tickets
      summary: Cancel Ticket
      operationId: cancel_ticket_api_tickets__ticket_id__cancel_post
      parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: string
          title: Ticket Id
      - name: X-Max-User-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Max-User-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/specialists:
    get:
      tags:
      - specialists
      summary: List Specialists
      operationId: list_specialists_api_specialists_get
      security:
      - HTTPBearer: []
      parameters:
      - name: organizationId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Organizationid
      - name: active_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Active Only
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpecialistOut'
                title: Response List Specialists Api Specialists Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - specialists
      summary: Create Specialist
      operationId: create_specialist_api_specialists_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpecialistIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecialistOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/specialists/{specialist_id}:
    patch:
      tags:
      - specialists
      summary: Update Specialist
      operationId: update_specialist_api_specialists__specialist_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: specialist_id
        in: path
        required: true
        schema:
          type: string
          title: Specialist Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpecialistUpdateIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecialistOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AnswerSnapshotOut:
      properties:
        questionId:
          type: string
          title: Questionid
        questionLabel:
          type: string
          title: Questionlabel
        optionId:
          type: string
          title: Optionid
        optionLabel:
          type: string
          title: Optionlabel
        optionCode:
          type: string
          title: Optioncode
      type: object
      required:
      - questionId
      - questionLabel
      - optionId
      - optionLabel
      - optionCode
      title: AnswerSnapshotOut
    AssignIn:
      properties:
        specialistId:
          type: string
          title: Specialistid
      type: object
      required:
      - specialistId
      title: AssignIn
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SpecialistIn:
      properties:
        organizationId:
          type: string
          title: Organizationid
        fullName:
          type: string
          title: Fullname
        skillTags:
          items:
            type: string
          type: array
          title: Skilltags
        active:
          type: boolean
          title: Active
          default: true
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
      type: object
      required:
      - organizationId
      - fullName
      title: SpecialistIn
    SpecialistOut:
      properties:
        id:
          type: string
          title: Id
        organizationId:
          type: string
          title: Organizationid
        fullName:
          type: string
          title: Fullname
        skillTags:
          items:
            type: string
          type: array
          title: Skilltags
        active:
          type: boolean
          title: Active
      type: object
      required:
      - id
      - organizationId
      - fullName
      - skillTags
      - active
      title: SpecialistOut
    SpecialistUpdateIn:
      properties:
        fullName:
          anyOf:
          - type: string
          - type: 'null'
          title: Fullname
        skillTags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Skilltags
        active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Active
      type: object
      title: SpecialistUpdateIn
    StatusHistoryOut:
      properties:
        status:
          $ref: '#/components/schemas/TicketStatus'
        at:
          type: string
          format: date-time
          title: At
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - status
      - at
      title: StatusHistoryOut
    SubmitTicketIn:
      properties:
        buildingId:
          type: string
          title: Buildingid
        categoryId:
          type: string
          title: Categoryid
        answers:
          additionalProperties:
            type: string
          type: object
          title: Answers
        photoUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Photourl
      type: object
      required:
      - buildingId
      - categoryId
      - answers
      title: SubmitTicketIn
    TicketOut:
      properties:
        id:
          type: string
          title: Id
        residentRef:
          type: string
          title: Residentref
        buildingId:
          type: string
          title: Buildingid
        addressSnapshot:
          type: string
          title: Addresssnapshot
        organizationId:
          type: string
          title: Organizationid
        parentCategoryId:
          type: string
          title: Parentcategoryid
        categoryId:
          type: string
          title: Categoryid
        answersSnapshot:
          items:
            $ref: '#/components/schemas/AnswerSnapshotOut'
          type: array
          title: Answerssnapshot
        recommendationTextSnapshot:
          type: string
          title: Recommendationtextsnapshot
        urgencyLevel:
          $ref: '#/components/schemas/UrgencyLevel'
        summaryText:
          type: string
          title: Summarytext
        status:
          $ref: '#/components/schemas/TicketStatus'
        photoUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Photourl
        assigneeSpecialistId:
          anyOf:
          - type: string
          - type: 'null'
          title: Assigneespecialistid
        takenByDispatcherId:
          anyOf:
          - type: string
          - type: 'null'
          title: Takenbydispatcherid
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
        statusHistory:
          items:
            $ref: '#/components/schemas/StatusHistoryOut'
          type: array
          title: Statushistory
      type: object
      required:
      - id
      - residentRef
      - buildingId
      - addressSnapshot
      - organizationId
      - parentCategoryId
      - categoryId
      - answersSnapshot
      - recommendationTextSnapshot
      - urgencyLevel
      - summaryText
      - status
      - createdAt
      - updatedAt
      - statusHistory
      title: TicketOut
    TicketStatus:
      type: string
      enum:
      - NEW
      - ACCEPTED
      - IN_PROGRESS
      - DONE
      - CANCELLED_BY_RESIDENT
      title: TicketStatus
    UrgencyLevel:
      type: string
      enum:
      - LOW
      - MEDIUM
      - HIGH
      title: UrgencyLevel
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer